notifications

+++ src/Modules/course/course.service.ts
@@ -4,6 +4,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Course, CourseContent, CourseDocument, CourseReview, Excercice, LessonType, random } from '../../Models/course.model';
 import { UserType } from '../../Models/user.model';
+import { Lang } from '../../shared/enums/lang.enum';
 import { OverrideUtils } from '../../shared/override-utils';
 import { CheckoutService } from '../checkout/checkout.service';
 import { UserService } from '../user/user.service';
@@ -49,17 +50,17 @@ export class CourseService {
         if (teacher['_id'].toString() !== req.user.id) {
             throw new BadRequestException('only teacher can update his courses');
         }
-        if (body.content){
-            body.content.forEach(element=>{
-                if (!element.OId) 
-                element.OId = OverrideUtils.generateGUID();
-                element.lessons.forEach(elem=>{
+        if (body.content) {
+            body.content.forEach(element => {
+                if (!element.OId)
+                    element.OId = OverrideUtils.generateGUID();
+                element.lessons.forEach(elem => {
                     elem.OId = OverrideUtils.generateGUID();
                 })
             })
         }
         await this.CourseModel.updateOne({ _id: id }, body).exec();
-        return await this.CourseModel.findById(id).exec();
+        return await this.findOne(req,id);
     }
 
     async delete(req: any, id: string): Promise<Course | PromiseLike<Course>> {
@@ -74,16 +75,10 @@ export class CourseService {
         if (course.startDate < Date.now()) {
             throw new BadRequestException('you can not delete started course');
         }
-        // let reservations = await this.checkoutService.CheckoutModel.find().populate({
-        //     "path": "lines.course",
-        //     'model': Course.name,
-        //     "match": new ObjectId(course['_id'].toString())
-        // }).populate('user')
-        // if (reservations.length>0){
-        //     throw new BadRequestException(`you have ${reservations.length}  reservation on your course you can't delete`);
-        // }
-
-
+        let orders = await this.checkoutService.CheckoutModel.exists({ course: new ObjectId(id) });
+        if (orders) {
+            throw new BadRequestException(req.user.defaultLang === Lang.en ? 'you can not delete course because you have reservations' : 'لا يمكن حذف الدورة لوجود حجوزات بها');
+        }
         return await this.CourseModel.findByIdAndDelete(id);
     }
 
@@ -124,19 +119,21 @@ export class CourseService {
     }
 
 
-    
+
     async findOne(req: any, id: string): Promise<Course | PromiseLike<Course>> {
         let course = await this.CourseModel.findById(id).exec();
 
         let reservations = await this.checkoutService.CheckoutModel.find({ course: new ObjectId(id) });
 
-        course.inCart = await this.userService.UserModel.exists({$and:[
-            { _id: new ObjectId(req.user.id)},
-            { cart: new ObjectId(id)},
-        ] })
-        console.log('inCart',course.inCart);
-        
-        course.purchased = await this.checkoutService.CheckoutModel.exists({ $and:[{course: new ObjectId(id)}, {user: new ObjectId(req.user.id)}] });
+        course.inCart = await this.userService.UserModel.exists({
+            $and: [
+                { _id: new ObjectId(req.user.id) },
+                { cart: new ObjectId(id) },
+            ]
+        })
+        console.log('inCart', course.inCart);
+
+        course.purchased = await this.checkoutService.CheckoutModel.exists({ $and: [{ course: new ObjectId(id) }, { user: new ObjectId(req.user.id) }] });
         let teacherCourses = await this.CourseModel.find({ teacher: course.teacher });
         course.teacher['cRating'] = teacherCourses.length > 0 ? teacherCourses.reduce((acc, course) => acc + course.cRating, 0) / teacherCourses?.length : 5;
         course.related = await this.CourseModel.find({
@@ -162,6 +159,7 @@ export class CourseService {
         }
         course.students = students
         course.enrolled = reservations.length;
+        course.progress = this.calculateProgress(course);
 
         course.related = course.related.slice(0, 6);
         return course;
@@ -189,6 +187,9 @@ export class CourseService {
         let courses = [];
         if (teacher) {
             courses = await this.CourseModel.find({ teacher: teacher['_id'] }).exec();
+            courses.forEach(course => {
+                course.progress = this.calculateProgress(course);
+            })
         }
         return courses;
     }
@@ -210,16 +211,22 @@ export class CourseService {
         courses = courses.filter(course => {
             let today = course.Days.find(day => OverrideUtils.dayOffDay(day) === dayOfWeek);
             let notFinished = course.content.find(content => content.lessons.find(lesson => lesson.isDone == false || lesson.isDone == null || lesson.isDone == undefined));
+
             return today && notFinished
         })
 
+        courses.forEach(course => {
+            course.progress = this.calculateProgress(course);
+        })
         return courses;
     }
 
     async getStudentCourses(req: any): Promise<Course[] | PromiseLike<Course[]>> {
         let purchased = await this.checkoutService.CheckoutModel.find({ user: new ObjectId(req.user.id) }).sort({ 'valueDate': 'desc' }).exec();
 
-
+        purchased.forEach(checkout => {
+            checkout.course.progress = this.calculateProgress(checkout.course);
+        })
         return purchased.map((checkout) => checkout.course);
     }
 
@@ -260,4 +267,20 @@ export class CourseService {
     }
 
 
+    /**
+     * calculateProgress
+     */
+    public calculateProgress(course: Course) {
+        let progress = 0;
+        let videos = 0;
+        course.content.forEach(cont => {
+            let contentProgress = 0;
+            cont.lessons.forEach(less => {
+                less.type.toString() === 'video' ? videos += 1 : videos += 0;
+                less.type.toString() === 'video' && less.isDone ? contentProgress += 1 : contentProgress += 0;
+            });
+            progress += contentProgress;
+        });
+        return videos === 0 ? 0 : progress / videos * 100;
+    }
 }
